Added Popochiu prefix to some base Objects.
Affected objects are Clickable, Prop, Hotspot, Region and DialogOption.
Removed references to Popochiu in some plugin's nodes so the Output does
not show all those annoying bugs when using the plugin for the first
time.
Removed the default CollisionPolygon2D from Character, Hotspot and Prop
so the nodes doesn't have to be set as Editable when adding them to a
scene.
=== Ready to work on allowing the plugin to remove Props and Hotspots
from a Room and on refactoring how Character animations are handled ===