Removed animation dependencies in PopochiuCharacter. (#3)
Now this Object only has a Sprite. Animations will be a responsability of developers.
Inspector allows to define how to flip the character Sprite when moving. By default is set to None.
Made a minor improvement in ICharacter so developers are not responsible of disabling the GI when calling `character_say` or `player_say`.
Replaced some code comments with English version.